BabyKit- Enable 3rdParty to use Calendar and NumericUpDown control in WeightPage
In win8 metro app, to reference a dll, it also need the xml outside of dll.
http://social.msdn.microsoft.com/Forums/windowsapps/en-US/51111470-8a86-44d4-acb8-e268afa7564e/payload-file-does-not-exist-compile-error?forum=winappswithcsharp
You're going to hate this - this is the ultimate solution.  Metro apps do not allow for embedded content, and that includes XAML pages.  Distributing classes is going to have to be done as a package of dlls and other resources, not a single dll.  Those responsible for this behavior have heard all of the issues around this, and it's not going to change.

Then it enable the weight page.